backend/requirements.txt
Django==5.2.4
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.16.0
psycopg2==2.9.10
pillow==11.3.0
pandas==2.3.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
tqdm==4.67.1
isort==6.0.1
setuptools==80.9.0
-e ./parse_invoice_letter
djangorestframework-simplejwt==5.5.1
pyjwt==2.10.1
python-decouple==3.8
gunicorn==25.1.0